+++ DataCollection/interface/RootEventInterface.hh
@@ -18,7 +18,7 @@ class RootEventInterface : public EventInterface
         ParticleCollection<Particle> getRecoParticles() const override;
         ParticleCollection<Particle> getRecoJets() const override;
         double getMET() const override;
-        long getEventIDNum() const override; 
+        unsigned long long getEventIDNum() const override; 
         std::vector<bool> getTriggerResults(std::string subProcess) const override;
         std::vector<std::string> getTriggerNames(std::string subProcess) const override;
         virtual int getNumPileUpInteractions() const override;
