+++ DataCollection/src/CmsswEventInterface.cc
@@ -113,7 +113,7 @@ double CmsswEventInterface::getMET() const
     throw std::runtime_error("There are no MET objects found");
 }
 
-long CmsswEventInterface::getEventIDNum() const
+unsigned long long CmsswEventInterface::getEventIDNum() const
 {
     throw std::runtime_error("Not implemented yet");
 }
