+++ EventFiles/interface/EventFile.hh
@@ -24,7 +24,7 @@ class EventFile
         virtual int getNumOfEvents() const = 0;
         int getEventCount() const {return eventCount;}
         virtual double getMET() const = 0;
-        virtual long getEventIDNum() const = 0; 
+        virtual unsigned long long getEventIDNum() const = 0; 
         virtual int getNumPileUpInteractions() const = 0;
         TFile* getFile() {return file;}
         virtual bool checkTrigger(std::string triggerName, std::string subProcess) const = 0; //checks if event passes trigger criteria
